faceRec/requirements.txt
torch==1.10.0
numpy==1.14.5
matplotlib==2.1.2
tqdm==4.23.4
#mxnet_cu90==1.2.1
scipy==1.0.0
bcolz==1.2.1
easydict==1.7
opencv_python==3.4.0.12
Pillow==5.2.0
mxnet==1.2.1.post1
scikit_learn==0.19.2
tensorboardX==1.2
torchvision==0.2.1
